# swagger-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Swingnotes",
    "version": "1.0.0",
    "description": "An api for Swingnotes"
  },
  "paths": {
    "/user/signup": {
      "post": {
        "summary": "Create a new user",
        "description": "Create a new user with the provided username, password, and email.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User created successfully"
          },
          "403": {
            "description": "Username already in use"
          },
          "406": {
            "description": "Validation error(s) or invalid email format"
          },
          "500": {
            "description": "Failed to create account"
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "summary": "User login",
        "description": "Log in a user with the provided username and password.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged in successfully"
          },
          "403": {
            "description": "Incorrect login credentials"
          },
          "500": {
            "description": "Failed to log in"
          }
        }
      }
    },
    "/notes": {
      "get": {
        "summary": "Get user notes",
        "description": "Retrieve all notes associated with the authenticated user.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Notes successfully retrieved"
          },
          "500": {
            "description": "Failed to retrieve notes"
          }
        }
      },
      "post": {
        "summary": "Create a new note",
        "description": "Create a new note with the provided title and text.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "text": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Note created successfully"
          },
          "406": {
            "description": "Validation error(s)"
          },
          "500": {
            "description": "Failed to create note"
          }
        }
      }
    },
    "/notes/{noteId}": {
      "put": {
        "summary": "Update a note",
        "description": "Update the title or text of a note.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "noteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the note to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "text": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Note updated successfully"
          },
          "403": {
            "description": "Unauthorized to update this note"
          },
          "404": {
            "description": "Note not found"
          },
          "500": {
            "description": "Failed to update note"
          }
        }
      },
      "delete": {
        "summary": "Delete a note",
        "description": "Delete a note by its ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "noteId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the note to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Note deleted successfully"
          },
          "403": {
            "description": "Unauthorized to delete this note"
          },
          "404": {
            "description": "Note not found"
          },
          "500": {
            "description": "Failed to delete note"
          }
        }
      }
    }
  }
}